Move auth files, separate signIn files, revert MemoryStorage usage, and other improvements
