+++ packages/agents-hosting/src/app/index.ts
@@ -3,7 +3,6 @@ export * from './agentApplicationBuilder'
 export * from './agentApplicationOptions'
 export * from './appRoute'
 export * from './attachmentDownloader'
-export * from './authorization'
 export * from './conversationUpdateEvents'
 export * from './routeHandler'
 export * from './routeList'
