+++ packages/agents-hosting/src/oauth/index.ts
@@ -1,3 +1,7 @@
 export * from './userTokenClient'
 export * from './oAuthFlow'
 export * from './userTokenClient.types'
+export * from './authorization'
+export * from './authorization.types'
+export * from './signInContext'
+export * from './signInStorage'
